+++ src/main/java/Qatar2022/tn/repository/equiperepository.java
@@ -1,9 +1,13 @@
 package Qatar2022.tn.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import Qatar2022.tn.entities.*;
+
+import Qatar2022.tn.entities.equipes;
+
+
 @Repository
-public interface equiperepository extends JpaRepository<equipes,Integer> {
+public interface equiperepository extends JpaRepository<equipes,Long> {
+
 
 }
